backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pandas==2.2.1
numpy==1.26.4
aif360==0.6.1
scikit-learn==1.4.1
google-generativeai==0.5.2
google-cloud-storage==2.16.0
google-cloud-bigquery==3.19.0
pydantic==2.6.4
python-dotenv==1.0.1
reportlab==4.1.0
fpdf2==2.7.9
httpx==0.27.0
